# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Revora Backend API
  version: "1.0.0"
  description: |
    Backend API for tokenized revenue-sharing on Stellar (offerings, investments,
    revenue reports, distributions, users, and notifications).
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/startup/offerings:
    get:
      summary: List offerings for startup
      description: |
        Returns offerings created by the authenticated startup user.
      tags: [Offerings]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Optional status filter.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 0
          description: Maximum number of offerings to return.
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Number of offerings to skip.
      responses:
        "200":
          description: List of offerings
          content:
            application/json:
              schema:
                type: object
                properties:
                  offerings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Offering"
                  total:
                    type: integer
                    description: Total number of matching offerings (if available).
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/offerings:
    get:
      summary: Public offerings catalog
      description: |
        Public catalog of offerings, accessible without authentication.
      tags: [Offerings]
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 0
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: sort
          schema:
            type: string
      responses:
        "200":
          description: List of public offerings
          content:
            application/json:
              schema:
                type: object
                properties:
                  offerings:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicOffering"
                  total:
                    type: integer
  /api/offerings/{id}:
    get:
      summary: Get offering by ID
      tags: [Offerings]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Offering details
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Offering"
                  - $ref: "#/components/schemas/PublicOffering"
        "400":
          description: Invalid ID format
        "404":
          description: Offering not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Offering:
      type: object
      properties:
        id:
          type: string
          format: uuid
        issuer_id:
          type: string
          description: Issuer user id
        contract_address:
          type: string
          description: On-chain contract address
        name:
          type: string
        symbol:
          type: string
        status:
          type: string
          description: Current status on/off chain
          enum: [draft, open, closed, paused, cancelled, active, completed]
        total_raised:
          type: string
          description: Total amount raised, as a high-precision decimal string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PublicOffering:
      type: object
      description: Public view of an offering
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        symbol:
          type: string
        status:
          type: string
        total_raised:
          type: string
        created_at:
          type: string
          format: date-time
...